change whitepaper link not to promote bitcoin.org, point to bitcoin.com instead.
Thanks to the Redditor who pointed out to me that the link to bitcoin.org promotes that site, which is run by folks who have resorted to censorship on forums they run.
That is certainly not something that BTCfork wants to endorse, so we change the link to bitcoin.com for now.